backend/requirements.txt
# FastAPI Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Data Processing
pandas==2.1.3
pyarrow==13.0.0

# Database
neo4j==5.15.0
sqlalchemy==2.0.23
pymysql==1.1.0

# LLM & API
requests==2.31.0
openai==1.3.8

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1

# Optional: For better async support
aiofiles==23.2.1
httpx==0.25.1
